+++ src/shared/apis/queryKeys.ts
@@ -1,3 +1,9 @@
+import type { ReviewListQueryParams } from '@/shared/types/review';
+
 export const queryKeys = {
-  REVIEWS: 'reviews',
+  REVIEWS: (params?: ReviewListQueryParams) => {
+    const rest = { ...params };
+    delete rest.page;
+    return ['reviews', Object.keys(rest).length ? rest : 'all'] as const;
+  },
 } as const;
