+++ src/widgets/review/api/useReviewListQuery.ts
@@ -1,5 +1,5 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
-import { fetchReviewList } from '@/shared/apis/review';
+import { fetchReviewList } from './fetchReviewList';
 import { queryKeys } from '@/shared/apis/queryKeys';
 import type { ReviewListQueryParams, ReviewListResponse } from '@/shared/types/review';
 
@@ -7,7 +7,7 @@ const INITIAL_PAGE = 1;
 
 export const useReviewListQuery = (params?: ReviewListQueryParams) =>
   useInfiniteQuery<ReviewListResponse>({
-    queryKey: params ? [queryKeys.REVIEWS, params] : [queryKeys.REVIEWS],
+    queryKey: queryKeys.REVIEWS(params),
     initialPageParam: INITIAL_PAGE,
     queryFn: ({ pageParam }) => {
       const currentPage = typeof pageParam === 'number' ? pageParam : INITIAL_PAGE;
